backend/requirements/dev.txt
asgiref==3.10.0
attrs==25.4.0
Django==4.2.25
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0

pytest
pytest-django
factory_boy
pytest-cov
